+++ nengo_edge/tests/test_saved_model_runner.py
@@ -13,7 +13,6 @@ from nengo_edge_models.asr.metrics import decode_predictions
 from nengo_edge_models.asr.models import lmuformer_tiny
 from nengo_edge_models.kws.models import lmu_small
 from nengo_edge_models.layers import Tokenizer
-from nengo_edge_models.models import MFCC
 from nengo_edge_models.models import Tokenizer as TokenizerDesc
 from nengo_edge_models.nlp.models import LMUformerEncoderNLP, lmuformer_small_sim
 
@@ -211,93 +210,6 @@ def test_nlp_tokenization(
     assert len(outputs) == len(gt)
     for x, y in zip(outputs, gt):
         np.testing.assert_allclose(x, y)
-
-
-def test_runner_streaming(
-    rng: np.random.RandomState, seed: int, tmp_path: Path
-) -> None:
-    tf.keras.utils.set_random_seed(seed)
-
-    interface = gpu.host.Interface(lmu_small(), build_dir=tmp_path)
-    assert isinstance(interface.pipeline.pre[0], MFCC)
-
-    # 3200 timesteps is equivalent to 200ms at 16 kHz
-    inputs = rng.uniform(-1, 1, size=(32, 3200)).astype("float32")
-    output0 = interface.run(inputs)
-
-    interface.export_model(tmp_path / "streaming", streaming=True)
-    runner = SavedModelRunner(tmp_path / "streaming")
-
-    # check that running in parts produces the same output
-    stream_chunk_size = inputs.shape[1] // 4
-    for i in range(4):
-        output1 = runner.run(
-            inputs[:, i * stream_chunk_size : (i + 1) * stream_chunk_size]
-        )
-
-    assert np.allclose(output0, output1, atol=1e-4), np.max(np.abs(output0 - output1))
-
-    # check that resetting state works
-    runner.reset_state()
-    for i in range(4):
-        output2 = runner.run(
-            inputs[:, i * stream_chunk_size : (i + 1) * stream_chunk_size]
-        )
-
-    assert np.allclose(output0, output2, atol=1e-4), np.max(np.abs(output0 - output2))
-
-    # test zero padding
-    runner.reset_state()
-    pad_output = runner.run(inputs[:, :10])
-    pad_output_gt = interface.run(
-        np.concatenate(
-            [
-                inputs[:, :10],
-                np.zeros((32, interface.pipeline.pre[0].window_size_samples - 10)),
-            ],
-            axis=1,
-        )
-    )
-    assert np.allclose(pad_output, pad_output_gt, atol=2e-6), np.max(
-        np.abs(pad_output - pad_output_gt)
-    )
-    pad_output_step = runner.run(inputs[:, -10:])
-    pad_output_step_gt = interface.run(
-        np.concatenate(
-            [
-                inputs[:, :10],
-                np.zeros((32, interface.pipeline.pre[0].window_size_samples - 10)),
-                inputs[:, -10:],
-                np.zeros((32, interface.pipeline.pre[0].window_stride_samples - 10)),
-            ],
-            axis=1,
-        )
-    )
-    assert np.allclose(pad_output_step, pad_output_step_gt, atol=2e-6), np.max(
-        np.abs(pad_output_step - pad_output_step_gt)
-    )
-
-
-def test_runner_state(tmp_path: Path) -> None:
-    pipeline = lmu_small()
-    interface = gpu.host.Interface(pipeline, build_dir=tmp_path)
-    interface.export_model(tmp_path, streaming=True)
-
-    runner = SavedModelRunner(tmp_path)
-
-    assert runner.state is None
-    runner.run(np.ones((3, 100)))
-    assert runner.state is not None
-    assert len(runner.state) == 9
-
-    with pytest.raises(ValueError, match="does not match saved state batch size"):
-        runner.run(np.ones((5, 100)))
-
-    runner.reset_state()
-    assert runner.state is None
-    runner.run(np.ones((5, 100)))
-    assert runner.state is not None
-    assert len(runner.state) == 9
 
 
 def test_runner_error_warnings(tmp_path: Path) -> None: